requirements-dev.txt
codacy-coverage
pytest<4.1  # see https://github.com/pytest-dev/pytest-cov/issues/252
pytest-cov>=2.6.1,<3
pytest-asyncio
flask>=1,<2
flake8
ipdb==0.10.3
factory_boy<3,>=2.9.2
psycopg2-binary>=2.7,<3
sqlalchemy>=1.3,<2
sqlalchemy_utils>=0.32.21,<1
attrs==19.1.0
